NCI-Agency/anet#2305: Change to a generic maintenance command
