+++ build.gradle
@@ -423,15 +423,25 @@ task integrationTest(type: Test) {
 	environment(run.environment)
 }
 
-// Create the task that runs the migration script for empty HTML biography fields of persons
-task dbClearEmptyBiographies(dependsOn: 'compileJava', type: JavaExec) {
-    group = "database"
-    description = "Runs the ANET clearEmptyBiographies command."
-    classpath = sourceSets.main.runtimeClasspath
-    environment(run.environment)
-    main = mainClassName
-    args = ["clearEmptyBiographies", anetConfig]
-    standardInput = System.in
+// Create the task that runs the maintenance command. Run e.g. with:
+//   ./gradlew dbMaintenance -Pargs="--clearEmptyBiographies"
+// To see all options, run just:
+//   ./gradlew dbMaintenance
+// which will print the help.
+task dbMaintenance(dependsOn: 'compileJava', type: JavaExec) {
+	group = "database"
+	description = "Runs the ANET database maintenance command."
+	classpath = sourceSets.main.runtimeClasspath
+	environment(run.environment)
+	main = mainClassName
+	def cmdline = ["maintenance"]
+	if (project.hasProperty("args")) {
+		cmdline.addAll(project.args.split("\\s+"))
+	} else {
+		cmdline << "--help"
+	}
+	cmdline << anetConfig
+	args cmdline
 }
 
 classes.dependsOn buildClient
