+++ src/main/java/mil/dds/anet/AnetApplication.java
@@ -98,7 +98,6 @@ public class AnetApplication extends Application<AnetConfiguration> {
   protected void addDefaultCommands(Bootstrap<AnetConfiguration> bootstrap) {
     bootstrap.addCommand(new ServerCommand<>(this));
     bootstrap.addCommand(new AnetCheckCommand(this));
-    bootstrap.addCommand(new ClearEmptyBiographiesCommand(this));
   }
 
   @Override
@@ -124,6 +123,9 @@ public class AnetApplication extends Application<AnetConfiguration> {
 
     // Add the database script command
     bootstrap.addCommand(new DatabaseScriptCommand());
+
+    // Add the database maintenance command
+    bootstrap.addCommand(new MaintenanceCommand(this));
 
     // Serve assets on /assets
     bootstrap.addBundle(new ConfiguredAssetsBundle(ImmutableMap.<String, String>builder()
